Build: test detection patterns
Do not rely on ASM to scan for test classes anymore.
Specify pattern instead.
This is mandatory for JDK 8 support until Gradle use ASM 5.

This commit save the gradle snippet and make the actual build use it already.
It was done in an effort to build QiWeb with JDK8.
It is a dead end for now as Gradle needs to use ASM5 for compileGroovy.
Same thing for Jacoco that use ASM under the hood.